Extract highest-passed-level reducer for reuse
Pull the "highest difficulty_level passed, per activity_type" loop out
of computeStudentTitles into a standalone highestPassedLevelByType in
sessionRules.ts, so the session-start route can reuse the exact same
reduction instead of reimplementing it.
